Password system
Implemented secure password System with a salted Hash. (NOTE: Prelim
password is "asdfawef".) It is possible to change the password itself to
something you want in case you change the password and then forget. Go
to src/data/settings.config, change the bottom line to "test" and then
let the top line be the md5 hash of "Insert Password Here" + "test". Of
course, this is insecure, but the student will not have enough time to
change the password when signing in without arousing suspicion, even
assuming they had the knowledge.